% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{List of amino acids.}
\format{
A data frame with 22 rows and 12 variables
}
\usage{
amino_acids
}
\description{
This dataset contains the list and characteristics of
amino acids.
}
\keyword{datasets}
