% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis_dm.R
\name{asis_dm}
\alias{asis_dm}
\title{asis_dm}
\usage{
asis_dm(composition, dm, digits = 2)
}
\arguments{
\item{composition}{feed chemical composition (as-is basis).}

\item{dm}{percentage of dry matter contents (without moisture) of feed.}

\item{digits}{integer indicating the number of decimal places}
}
\description{
This function convert as-is basis into dry-matter basis.
}
\examples{
df <- data.frame(nutrient=c("CP","NDF","EE","Ash"),composition=c(10,30,5,2))
asis_dm(df$composition,85)
}
\keyword{feed}
\keyword{formulation}
