% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_j.R
\name{cal_j}
\alias{cal_j}
\title{cal_j}
\usage{
cal_j(cal, standard = "15")
}
\arguments{
\item{cal}{calorie.}

\item{standard}{"thermochemical" = Thermochemical calorie; "4" = 4°C calorie; "15" = 15°C calorie; "20" = 20°C calorie; "mean" = mean calorie (https://en.wikipedia.org/wiki/Calorie).}
}
\description{
This function converts calorie into joule.
}
\examples{
cal_j(cal = 5000, standard = "15")
}
\keyword{feed}
\keyword{formulation}
