% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatty_acids}
\alias{fatty_acids}
\title{List of fatty acids.}
\format{
A list with 2 tibbles (saturated & unsaturated)
}
\usage{
fatty_acids
}
\description{
This dataset contains the list and characteristics of
fatty acids.
}
\keyword{datasets}
