% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_compare.R
\name{feed_compare}
\alias{feed_compare}
\title{feed_compare}
\usage{
feed_compare(feed1, feed2, name1 = "first", name2 = "second",
  radar = TRUE)
}
\arguments{
\item{feed1}{1st feed chemical composition that you want to compare.}

\item{feed2}{2nd feed chemical composition that you want to compare.}

\item{name1}{1st feed name that you want to compare.}

\item{name2}{2nd feed name that you want to compare.}

\item{radar}{TRUE returns radar chart.}
}
\description{
A function to compare each result of feed composition.
}
\examples{
feed1 <- feed::feed_info("https://www.feedipedia.org/node/556", 2)
feed2 <- feed::feed_info("https://www.feedipedia.org/node/674", 3)
feed1 <- feed1[[1]]
feed2 <- feed2[[1]]
feed_compare(feed1, feed2, name1 = "corn", name2 = "sbm", radar = TRUE)
}
\keyword{feed}
\keyword{formulation}
