% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_info.R
\name{feed_info}
\alias{feed_info}
\title{feed_info}
\usage{
feed_info(url)
}
\arguments{
\item{url}{A URL that you want to scraping.}
}
\description{
This function show the result of tables of URL.
}
\examples{
feed_info(url="https://www.feedipedia.org/node/556")
}
\keyword{feed,}
\keyword{feedipedia}
