% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_list.R
\name{feed_list}
\alias{feed_list}
\title{feed_list}
\usage{
feed_list(category = "all")
}
\arguments{
\item{category}{A category that you want to find(all; forage; plantProduct; animalOrigin; other)}
}
\description{
This function showed a feed list at feedipedia.org.
}
\examples{
feed_list("all")
feed_list("forage")
feed_list("plantProduct")
feed_list("animalOrigin")
feed_list("other")
}
\keyword{feed,}
\keyword{feedipedia}
