% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feed_nias}
\alias{feed_nias}
\title{Feed ingredient table of NIAS.}
\format{A data frame with XX rows and XX variables}
\usage{
feed_nias
}
\description{
This dataset contains a chemical composition of
feed ingredient of National Institute of Animal Science.
}
\keyword{datasets}
