% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_search.R
\name{feed_search}
\alias{feed_search}
\title{feed_search}
\usage{
feed_search(keyword)
}
\arguments{
\item{keyword}{A keyword that you want to find.}
}
\description{
This function search a keyword at feedipedia.org.
}
\examples{
feed_search("alfalfa")
}
\keyword{feed,}
\keyword{feedipedia}
