% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrc_beef.R
\docType{data}
\name{nrc_beef}
\alias{nrc_beef}
\title{nrc_beef}
\format{A data frame with 219 rows and 57 variables:}
\usage{
nrc_beef
}
\description{
Feed chemical composition from NRC beef 2016
}
\keyword{datasets}
