% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdn.R
\name{tdn}
\alias{tdn}
\title{tdn}
\usage{
tdn(cp, ee, cf, nfe)
}
\arguments{
\item{cp}{CP.}

\item{ee}{EE.}

\item{cf}{CF.}

\item{nfe}{NFE(100 - CP - EE - CF - Ash).}
}
\description{
This function calculate the TDN of cattle. Equation is TDN = 0.85CP + 2.25*0.8EE + 0.6CF + 0.8NFE.
}
\examples{
tdn(cp = 18, ee = 9, cf = 35, nfe = 19)
}
\keyword{feed}
\keyword{formulation}
