% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_info.R
\name{feed_info}
\alias{feed_info}
\title{feed_info}
\usage{
feed_info(url, extract = 0)
}
\arguments{
\item{url}{A URL that you want to scraping.}

\item{extract}{A number of list for nutreint composition.}
}
\description{
This function show the result of tables of URL. If you use extract option, a list is returned for each nutrient composition of the specific feed.
}
\examples{
feed_info(url="https://www.feedipedia.org/node/556")
feed_info(url="https://www.feedipedia.org/node/556",extract=5)
}
\keyword{feed,}
\keyword{feedipedia}
